+++ Pluto/IO/Binary/StrictStreamBinaryReader.cs
@@ -9,5 +9,5 @@ public class StrictStreamBinaryReader : StreamBinaryReader {
 	public StrictStreamBinaryReader(string path, FileAccess access = FileAccess.Read, FileShare share = FileShare.ReadWrite, bool leaveOpen = false) : this(new FileStream(path, FileMode.Open, access, share), leaveOpen) { }
 	public StrictStreamBinaryReader(FileInfo fileInfo, FileAccess access = FileAccess.Read, FileShare share = FileShare.ReadWrite, bool leaveOpen = false) : this(new FileStream(fileInfo.FullName, FileMode.Open, access, share), leaveOpen) { }
 
-	public override int Position { get => (int) BaseStream.Position; set => BaseStream.Position = value; }
+	public override long Position { get => (int) BaseStream.Position; set => BaseStream.Position = value; }
 }
